Refactor shared types into separate directory
Even though these shared structures are currently only used on the
server, the client UI will ultimately need some of these structures, so
we are creating a clear API for both the client and server.
