+++ src/plugins/profiling/common/profiling.ts
